remove generic footer + edit & use waitlist footer + shrink page top padding
